refactor: consolidate value provider tests and remove placeholder pattern
- Removed placeholder values from DaprDistributedApplicationLifecycleHook
- Replaced with cleaner hasValueProviders flag for tracking value providers
- Consolidated tests from EndpointReferenceTests into ResourceBuilderExtensionsTests
- Deleted redundant EndpointReferenceTests.cs file
- All tests pass with cleaner separation of concerns
